requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.1
Django==4.2.2
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.0.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
